// src/models/loginForms.ts
/**
 * Created by hale on 2016/11/14.
 */

export class loginForms {
    phone: string;
    smsCode: string;
    encryptSmsCode: string;
    errorMsg: string;
    errorFlag: boolean;
    account: string;
    password: string;
    clickFlag: number;
    stateType: string;
    btnSendCodeText: string;
    sec: number;
    valid: boolean;
}